% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbiwc.R
\name{dbiWC}
\alias{dbiWC}
\title{DBI-WC Clustering Algorithm}
\usage{
dbiWC(sam, k, max_itr = 20, tol = 1e-04)
}
\arguments{
\item{sam}{List of matrices where each matrix represents a sample observation}

\item{k}{Integer number of clusters}

\item{max_itr}{Maximum number of iterations (default: 20)}

\item{tol}{Convergence tolerance (default: 1e-4)}
}
\value{
List containing:
\tabular{ll}{
\code{"label"} \tab Cluster assignments. \cr
\code{"mdbi"} \tab Davies-Bouldin Index values for each feature. \cr
\code{"weight"} \tab Learned distance weights. \cr
\code{"center"} \tab Final cluster centers. \cr
\code{"num_itr"} \tab Number of iterations performed. \cr
}
}
\description{
The DBI-WC clustering algorithm is a novel method designed for mixed-type distribution data. It employs a modified Davies-Bouldin Index (DBI) to construct a DBI-weighted Wasserstein distance metric, extending the traditional Wasserstein distance. Combined with a robust initialization strategy, DBI-WC achieves strong performance on mixed-type distribution data.
}
\details{
DBI-WC Clustering Algorithm
}
