% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbiwc.R
\name{saWC}
\alias{saWC}
\title{SA-WC Clustering Algorithm}
\usage{
saWC(sam, k, max_itr = 20, tol = 1e-04)
}
\arguments{
\item{sam}{List of matrices where each matrix represents a sample observation}

\item{k}{Integer number of clusters}

\item{max_itr}{Maximum number of iterations (default: 20)}

\item{tol}{Convergence tolerance (default: 1e-4)}
}
\value{
List containing:
\tabular{ll}{
\code{"label"} \tab Cluster assignments. \cr
\code{"weight"} \tab Learned distance weights. \cr
\code{"center"} \tab Final cluster centers. \cr
\code{"num_itr"} \tab Number of iterations performed. \cr
}
}
\description{
The SA-WC algorithm is a simplified version of DBI-WC. It uses a straightforward weighting scheme to integrate various distribution data. All other specifications are consistent with those of DBI-WC.
}
